Conform with -Wsign-conversion, part 1
Fixes -Wsign-conversion warnings in common/, replay/, samples/,
utils/, and native/ files that aren't from a specific backend.

Also suppresses the warning in native/opengl/ because the GL API
forces an excessive number of sign casts to no apparent benefit.
tests/ are also suppressed since it requires a large number of
fixes but does not contribute to hardening Dawn.

-Wsign-conversion is not enabled in this CL, since there are other
areas of the code that still need to be updated prior to adding the
flag.

Bug: 42240462
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/321275
Commit-Queue: Brandon Jones <bajones@chromium.org>

+++ src/dawn/samples/ManualSurfaceTest.cpp
@@ -196,8 +196,8 @@ void SyncFromWindow(WindowData* data) {
     int height;
     glfwGetFramebufferSize(data->window, &width, &height);
 
-    data->targetConfig.width = std::max(1u, width / data->divisor);
-    data->targetConfig.height = std::max(1u, height / data->divisor);
+    data->targetConfig.width = std::max(1u, static_cast<uint32_t>(width) / data->divisor);
+    data->targetConfig.height = std::max(1u, static_cast<uint32_t>(height) / data->divisor);
 }
 
 void AddWindow() {
